+++ Back-End/WaltCommerce.Api/DataLayer/ContextExtension/RoleConfiguration.cs
@@ -1,0 +1,15 @@
+﻿using Common.Models.Roles;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace DataLayer.ContextExtension
+{
+    public class RoleConfiguration : IEntityTypeConfiguration<Role>
+    {
+        public void Configure(EntityTypeBuilder<Role> builder)
+        {
+            //TODO set Name Unique
+            builder.Property(x => x.Name).IsRequired().HasMaxLength(10);
+        }
+    }
+}
