--- Back-End/WaltCommerce.Api/DataLayer/ContextExtension/Roles/RoleConfiguration.cs
@@ -1,15 +1,0 @@
-﻿using Common.Models.Roles;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace DataLayer.ContextExtension.Roles
-{
-    public class RoleConfiguration : IEntityTypeConfiguration<Role>
-    {
-        public void Configure(EntityTypeBuilder<Role> builder)
-        {
-            //TODO set Name Unique
-            builder.Property(x => x.Name).IsRequired().HasMaxLength(10);
-        }
-    }
-}
