--- Back-End/WaltCommerce.Api/DataLayer/ContextExtension/UserAddresses/UserAddressesConfiguration.cs
@@ -1,14 +1,0 @@
-﻿using Common.Models.UsersAddresses;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace DataLayer.ContextExtension.UserAddresses
-{
-    class UserAddressesConfiguration : IEntityTypeConfiguration<UserAddress>
-    {
-        public void Configure(EntityTypeBuilder<UserAddress> builder)
-        {
-            builder.Property(x => x.Id).IsRequired().UseIdentityColumn().UseSerialColumn();
-        }
-    }
-}
