+++ Back-End/WaltCommerce.Api/DataLayer/Contexts/MainDbContext.cs
@@ -9,22 +9,10 @@ using Common.Models.ProductsSpecifications;
 using Common.Models.Roles;
 using Common.Models.Sizes;
 using Common.Models.Users;
-using DataLayer.ContextExtension.Brands;
-using DataLayer.ContextExtension.BuyCartDetails;
-using DataLayer.ContextExtension.BuyCarts;
-using DataLayer.ContextExtension.PersonTypeCategories;
-using DataLayer.ContextExtension.ProductCategories;
-using DataLayer.ContextExtension.ProductColors;
-using DataLayer.ContextExtension.Products;
-using DataLayer.ContextExtension.ProductsImages;
-using DataLayer.ContextExtension.ProductsSizes;
-using DataLayer.ContextExtension.ProductsSpecifications;
-using DataLayer.ContextExtension.Roles;
-using DataLayer.ContextExtension.Users;
+using DataLayer.ContextExtension;
 using Common.Models.Products;
 using Microsoft.EntityFrameworkCore;
 using Common.Models.UsersAddresses;
-using DataLayer.ContextExtension.UserAddresses;
 using Common.Models.Orders;
 using Common.Models.OrderDetails;
